Fixed colons in filenames confusing -D
The hexdigest length is now included as the first line in the output of
--checksum, and is read when performing a diff with -D to determine how
much to read from each line until the file path begins. Ideally, no
number would be needed, but since Sha2 support is a must, varying
lengths are a thing. A more elegant way would be to default to Xxh3's
length unless an algorithm is provided in conjunction with -D. Will
implement that in a future commit.